Unnecessary casts (parameters already of correct type)
git-svn-id: https://svn.apache.org/repos/asf/commons/proper/math/trunk@1206655 13f79535-47bb-0310-9956-ffa450edef68

+++ src/main/java/org/apache/commons/math/util/MathUtils.java
@@ -209,7 +209,7 @@ public final class MathUtils {
                    magnitude == Integer.MIN_VALUE) {
             throw new MathArithmeticException(LocalizedFormats.OVERFLOW);
         } else {
-            return (int) -magnitude; // Flip sign.
+            return -magnitude; // Flip sign.
         }
     }
 
@@ -231,7 +231,7 @@ public final class MathUtils {
                    magnitude == Long.MIN_VALUE) {
             throw new MathArithmeticException(LocalizedFormats.OVERFLOW);
         } else {
-            return (long) -magnitude; // Flip sign.
+            return -magnitude; // Flip sign.
         }
     }
